--- a/src/Microsoft.AspNet.Server.Kestrel/Infrastructure/IStringCache.cs
+++ b/src/Microsoft.AspNet.Server.Kestrel/Infrastructure/IStringCache.cs
@@ -1,0 +1,8 @@
+﻿namespace Microsoft.AspNet.Server.Kestrel.Infrastructure
+{
+    public interface IStringCache
+    {
+        void MarkStart();
+        unsafe string GetString(char* data, uint hash, int length);
+    }
+}