--- a/src/Microsoft.AspNet.Server.Kestrel/KestrelServer.cs
+++ b/src/Microsoft.AspNet.Server.Kestrel/KestrelServer.cs
@@ -60,7 +60,15 @@ namespace Microsoft.AspNet.Server.Kestrel
                 {
                     FrameFactory = (context, remoteEP, localEP, prepareRequest) => 
                     {
-                        return new Frame<TContext>(application, context, remoteEP, localEP, prepareRequest);
+                        return new Frame<TContext>(
+                            application, 
+                            context, 
+                            remoteEP, 
+                            localEP, 
+                            prepareRequest, 
+                            information.StringCacheOnConnection ? 
+                                new StringCache(information.StringCacheMaxStrings, information.StringCacheMaxStringLength) : 
+                                null);
                     },
                     AppLifetime = _applicationLifetime,
                     Log = trace,