--- a/src/Microsoft.AspNet.Server.Kestrel/KestrelServerInformation.cs
+++ b/src/Microsoft.AspNet.Server.Kestrel/KestrelServerInformation.cs
@@ -16,6 +16,8 @@ namespace Microsoft.AspNet.Server.Kestrel
             Addresses = GetAddresses(configuration);
             ThreadCount = threadCount;
             NoDelay = true;
+
+            ConfigureStringCache(configuration);
         }
 
         public ICollection<string> Addresses { get; }
@@ -23,6 +25,12 @@ namespace Microsoft.AspNet.Server.Kestrel
         public int ThreadCount { get; set; }
 
         public bool NoDelay { get; set; }
+
+        public bool StringCacheOnConnection { get; set; }
+
+        public int StringCacheMaxStrings { get; set; }
+
+        public int StringCacheMaxStringLength { get; set; }
 
         public IConnectionFilter ConnectionFilter { get; set; }
 
@@ -38,6 +46,49 @@ namespace Microsoft.AspNet.Server.Kestrel
             }
 
             return addresses;
+        }
+
+        private void ConfigureStringCache(IConfiguration configuration)
+        {
+            bool stringCacheOnConnection;
+            if (bool.TryParse(configuration["server.stringCacheOnConnection"], out stringCacheOnConnection))
+            {
+                StringCacheOnConnection = stringCacheOnConnection;
+            }
+            else
+            {
+                StringCacheOnConnection = true;
+            }
+            int stringCacheMaxStrings;
+            if (StringCacheOnConnection && int.TryParse(configuration["server.stringCacheMaxStrings"], out stringCacheMaxStrings))
+            {
+                if (stringCacheMaxStrings <= 0)
+                {
+                    throw new ArgumentOutOfRangeException(nameof(stringCacheMaxStrings),
+                        stringCacheMaxStrings,
+                        "StringCacheMaxStrings must be positive.");
+                }
+                StringCacheMaxStrings = stringCacheMaxStrings;
+            }
+            else
+            {
+                StringCacheMaxStrings = 25;
+            }
+            int stringCacheMaxStringLength;
+            if (StringCacheOnConnection && int.TryParse(configuration["server.stringCacheMaxStringLength"], out stringCacheMaxStringLength))
+            {
+                if (stringCacheMaxStringLength <= 0)
+                {
+                    throw new ArgumentOutOfRangeException(nameof(stringCacheMaxStringLength),
+                        stringCacheMaxStringLength,
+                        "StringCacheMaxStringLength must be positive.");
+                }
+                StringCacheMaxStringLength = stringCacheMaxStringLength;
+            }
+            else
+            {
+                StringCacheMaxStringLength = 256;
+            }
         }
     }
 }