--- a/test/Microsoft.AspNet.Server.KestrelTests/AsciiDecoder.cs
+++ b/test/Microsoft.AspNet.Server.KestrelTests/AsciiDecoder.cs
@@ -13,6 +13,7 @@ namespace Microsoft.AspNet.Server.KestrelTests
         [Fact]
         private void FullByteRangeSupported()
         {
+            var stringCache = new StringCache();
             var byteRange = Enumerable.Range(0, 255).Select(x => (byte)x).ToArray();
 
             var mem = MemoryPoolBlock2.Create(new ArraySegment<byte>(byteRange), IntPtr.Zero, null, null);
@@ -21,7 +22,7 @@ namespace Microsoft.AspNet.Server.KestrelTests
             var begin = mem.GetIterator();
             var end = GetIterator(begin, byteRange.Length);
 
-            var s = begin.GetAsciiString(end);
+            var s = begin.GetAsciiString(end, stringCache);
 
             Assert.Equal(s.Length, byteRange.Length);
 
@@ -37,6 +38,7 @@ namespace Microsoft.AspNet.Server.KestrelTests
         [Fact]
         private void MultiBlockProducesCorrectResults()
         {
+            var stringCache = new StringCache();
             var byteRange = Enumerable.Range(0, 512 + 64).Select(x => (byte)x).ToArray();
             var expectedByteRange = byteRange
                                     .Concat(byteRange)
@@ -60,7 +62,7 @@ namespace Microsoft.AspNet.Server.KestrelTests
             var begin = mem0.GetIterator();
             var end = GetIterator(begin, expectedByteRange.Length);
 
-            var s = begin.GetAsciiString(end);
+            var s = begin.GetAsciiString(end, stringCache);
 
             Assert.Equal(s.Length, expectedByteRange.Length);
 
@@ -76,6 +78,7 @@ namespace Microsoft.AspNet.Server.KestrelTests
         [Fact]
         private void HeapAllocationProducesCorrectResults()
         {
+            var stringCache = new StringCache();
             var byteRange = Enumerable.Range(0, 16384 + 64).Select(x => (byte)x).ToArray();
             var expectedByteRange = byteRange.Concat(byteRange).ToArray();
 
@@ -89,7 +92,7 @@ namespace Microsoft.AspNet.Server.KestrelTests
             var begin = mem0.GetIterator();
             var end = GetIterator(begin, expectedByteRange.Length);
 
-            var s = begin.GetAsciiString(end);
+            var s = begin.GetAsciiString(end, stringCache);
 
             Assert.Equal(s.Length, expectedByteRange.Length);