On a keep-alive connection most of the request header's *values* are the same (method, host, http version, cookie, user agent, language, path etc); since this string creation is the highest allocation over a connection, and generally the same per request, reuse the previous created string rather than creating new ones.

*Updated*

This change **cuts 88%** of allocs and **92.5%** of alloc'd memory for strings (and 33% of *total* allocations)

With current "aspnet/dev" + this; 4k `/plaintext?N` browser requests from Firefox pipelined 16 deep (where N = 1...200) e.g. 20 refreshes.

Before % allocations

![before 42% allocations](http://aoa.blob.core.windows.net/aspnet/string-pool-2-before.png)

After % allocations

![after 8% allocations](http://aoa.blob.core.windows.net/aspnet/string-pool-2-after.png)

Before total allocations: 44,112 strings, 2.8 MBytes

![before 2,804 kB allocations](http://aoa.blob.core.windows.net/aspnet/string-pool-2-before-totals.png)

After total allocations: 5,619 string, 210 kBytes

![after 210 kB allocations](http://aoa.blob.core.windows.net/aspnet/string-pool-2-after-totals.png)

Resolves final part of #291