@benaadams 2015-12-02
From @halter73 on  #441
> @benaadams After discussing #411 with @davidfowl and @cesarbs we were thinking about closing it due to the relative complexity of managing a per-connection invalidating cache to reduce string allocations.
>
> I didn't what to start that discussion before offering an alternative, and this is it. I prefer this approach, because it has a better (or at least easier to analyze) worst case performance both memory and cpu-wise. This might also have a better memory footprint per-connection since we don't have to reallocate a new cache each time.