@benaadams 2015-12-02
Will rebase post #441, as can slim the buffers with Mehod and Version removed. I'll come back with more evidence either way (good or bad) after that.