@benaadams 2015-12-04
Hopefully this analysis is correct... might not be; errors and omissions accepted...

> worst case performance (...) memory

Looking at a single request, worse case overhead is same as best case as you've still allocated the strings an have them in the header collection; so its the 3 arrays which would be:

`25 * sizeof(unit) + 25 * sizeof(int) + 25 * sizeof(IntPtr) + 3 * array overhead`

On x64 `25 * 4 + 25 * 4 + 25 * 8 + 3 * 28` = 484 bytes per connection
On x86 `25 * 4 + 25 * 4 + 25 * 4 + 3 * 12` = 336 bytes per connection

Worse cast overhead by a bad actor; in strings would 25 strings of max length 256 * 25 * 2 = 12800 bytes + 25 * string overhead ; but without it you'd still allocate those strings for the headers anyway and maintain them for the request.

So probably real the worse case that deviates from without cache; would be if someone on keepalive was sending a 256 byte cookie that changed on every request when it would be (25 - connection repeated strings) * 256.

> cpu-wise

By itself the hashing should be more than x 2 faster than `String.GetHashCode` in [non-Marvin32](https://github.com/dotnet/coreclr/blob/master/src/mscorlib/src/System/String.cs#L801-L850) variant and simpler than `InternalMarvin32HashString` or  used by [Dictionary.Add](https://github.com/dotnet/coreclr/blob/master/src/mscorlib/src/System/Collections/Generic/Dictionary.cs#L339) as it operates on 4 chars(byte) per iteration rather than 2 chars (char) - however its also hashing at the same time as the string construction - so will be close(ish) to free.

So the main worse case will be a scan and compare of a contiguous array 25 `uint`s per unknown string which should occupy 2 cache lines - on match it does a x 4 parallel string compare to ensure it is actually the same string - a similar cost to a dictionary key lookup - though faster than a case insensitive look up.

Security-wise - Dictionary now uses `InternalMarvin32HashString` to deal DoS from predicted key collisions. I don't believe this has the same issue as its a bounded array of 25 elements; of maximum length 256. However; I do seed the hash with a random number on startup anyway; to prevent learnings from being transferable.

Having said all this... someone else should probably check it. Also I need to recheck its doing what it was doing before I changed the hashcode function to closely match the [span hashcode with fix](https://github.com/dotnet/corefxlab/pull/455) which had a lot of perf analysis done on it.