@benaadams 2015-12-04
@davidfowl would you hang on to the MemoryPoolBlocks from SocketInput until end of request? (header portion)