@benaadams 2015-12-04
Though using spans and Utf8String and not doing a copy at all to string would be better...