@davidfowl 2015-12-04
Ah this is *request* header values. My bad :smile: Can we just do more of what we @cesarbs change does? Intern a bunch of known things and do a fast comparison? This looks way too generic for what we want to do.