@halter73 2015-12-04
For some header values like "keep-alive", it is certainly possible do do something more like what @cesarbs's change does. I think it would get trickier for things like user agents. I'm not even sure things like "Accept*" header values are too consistent between browsers, regions, etc... Cookie values are obviously a no go.

It might not seem like cookie values would be reused too often (and that's likely true between different connections), but over a single connection, I expect cookie values get reused frequently.

@davidfowl In case you didn't notice, I'll point out that there is a unique StringPool per connection, so everything @benaadams bolded in his previous comment would likely successfully be cached for every subsequent request made over that connection.

Still, I do wonder about the real world impact of this change. Do we expect this will improve the performance of a typical high traffic website running on Kestrel? Or is this mostly just to help us out with our benchmarks? Or is there some other scenario where a lot of requests are made over a single connection that I'm not thinking about? Reverse proxies can reuse connections for a lot of requests from different users, but in that case I don't know if caching only the last 25 header values is sufficient.

To figure out if this change makes a difference for a typical web server, I imagine we would have to estimate how many requests are usually made over a single connection by a normal user browsing a website in Chrome, IE, etc..., and how many requests there would need to need to be for us to see a real performance improvement from this change. 

Unlike some previous performance improvements this isn't just a pure win. Obviously I can see some scenarios where this can reduce allocations quite a bit (like with our benchmarks), but there are clearly some scenarios (like with a single request per connection) where it wouldn't help at all. We could mitigate the single request thing by just not using the string pool for non-keep-alive connections, but that just covers one of the possible bad cases.

Long story short, I'm conflicted on this change, and when I'm unsure I like to error on the side of simplicity. I haven't closed this yet because maybe there's something I haven't considered that will change my thinking.