@halter73 2015-12-04
@benaadams To add to my previous comment, while a 33% reduction in total allocations is impressive, am I correct in understanding that's for 4k requests over a *single* connection? I'm also assuming the 16 deep pipelining is helping us reduce our SocketOutput related allocations by comparison.