@AspNetSmurfLab 2015-12-08
@benaadams I ran our plaintext benchmark scenario with the additional headers you suggested, and here are the results from our perf lab:

__Baseline:__

	wrk -c 64 -t 32 -d 60 -s ./scripts/pipeline.lua http://10.0.0.100:5001/plaintext -- 16
	Running 1m test @ http://10.0.0.100:5001/plaintext
	  32 threads and 64 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency     5.71ms   20.28ms 462.54ms   88.49%
	    Req/Sec    13.09k     2.56k   70.75k    71.27%
	  24999284 requests in 1.00m, 3.07GB read
	Requests/sec: 415961.30
	Transfer/sec:     52.36MB

	Allocated: 81.34 MB
	Gen 0: 174, Gen 1: 3, Gen 2: 2

	Running 1m test @ http://10.0.0.100:5001/plaintext
	  32 threads and 64 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency    10.19ms   33.73ms 520.07ms   88.47%
	    Req/Sec    12.64k     3.27k   60.91k    68.47%
	  24143877 requests in 1.00m, 2.97GB read
	Requests/sec: 401748.34
	Transfer/sec:     50.57MB

	Allocated: 52.68 MB
	Gen 0: 170, Gen 1: 3, Gen 2: 2

__With this PR merged:__

	Running 1m test @ http://10.0.0.100:5001/plaintext
	  32 threads and 64 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency     7.09ms   29.22ms 644.22ms   89.82%
	    Req/Sec    12.64k     3.15k   24.05k    69.16%
	  24136730 requests in 1.00m, 2.97GB read
	Requests/sec: 402109.14
	Transfer/sec:     50.62MB

	Allocated: 165.04 MB
	Gen 0: 165, Gen 1: 3, Gen 2: 2

	Running 1m test @ http://10.0.0.100:5001/plaintext
	  32 threads and 64 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency     6.88ms   22.68ms 500.36ms   88.23%
	    Req/Sec    13.09k     2.89k   63.51k    68.78%
	  25011742 requests in 1.00m, 3.07GB read
	Requests/sec: 416176.82
	Transfer/sec:     52.39MB
	
	Allocated: 214.21 MB
	Gen 0: 177, Gen 1: 3, Gen 2: 2

Note that "Allocated: " is the result `GC.GetTotalMemory(forceFullCollection: false)` about 15 seconds after the test is completed. So this isn't the amount of memory allocated during the test, but instead the amount of total active memory in the heap after the test is over. If I forced GC, I suspect the allocation numbers would be more similar between the baseline and the PR.

The main point of showing the GC numbers is to show that there were plenty of Gen 0 collections in each run.

All said, I don't think these results show a significant enough change in performance to warrant a string cache for headers. If you have any suggestions on parameters I should tweak when running these tests, I'm all ears.