@AspNetSmurfLab 2015-12-08
I did some more runs with more connections, but less time per run. I also didn't restart the server between runs, but I did do a warmup run before measuring. Here are the results:

__Baseline:__

	wrk -c 256 -t 32 -d 30 -s ./scripts/pipeline.lua http://10.0.0.100:5001/plaintext -- 16
	Running 30s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency    14.32ms   33.04ms 408.59ms   92.83%
	    Req/Sec    13.91k     1.79k   96.00k    92.22%
	  13284799 requests in 30.10s, 1.63GB read
	Requests/sec: 441356.37
	Transfer/sec:     55.56MB

	Running 30s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency    24.05ms   53.10ms 584.39ms   90.42%
	    Req/Sec    14.13k     2.26k  127.33k    86.96%
	  13509384 requests in 30.10s, 1.66GB read
	Requests/sec: 448862.90
	Transfer/sec:     56.51MB

	Running 30s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency    16.76ms   37.96ms 455.98ms   91.40%
	    Req/Sec    14.04k     1.80k   57.01k    88.76%
	  13399709 requests in 30.10s, 1.65GB read
	Requests/sec: 445189.55
	Transfer/sec:     56.04MB

	Allocated: 174.88 MB
	Gen 0: 378, Gen 1: 30, Gen 2: 4

	Forcing GC... done!

	Allocated: 21.48 MB
	Gen 0: 379, Gen 1: 31, Gen 2: 5

__With this PR merged:__

	Running 30s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency    15.54ms   33.68ms 407.07ms   91.68%
	    Req/Sec    13.99k     1.83k  152.34k    90.59%
	  13349015 requests in 30.10s, 1.64GB read
	  Socket errors: connect 0, read 0, write 300, timeout 0
	Requests/sec: 443499.07
	Transfer/sec:     55.83MB

	Running 30s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency    18.50ms   41.38ms 468.02ms   91.44%
	    Req/Sec    14.08k     1.75k   86.40k    86.57%
	  13462037 requests in 30.10s, 1.65GB read
	Requests/sec: 447262.28
	Transfer/sec:     56.30MB

	Running 30s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency    23.63ms   58.09ms 868.69ms   91.74%
	    Req/Sec    14.11k     1.76k   57.84k    81.93%
	  13498453 requests in 30.10s, 1.66GB read
	Requests/sec: 448466.09
	Transfer/sec:     56.46MB

	Allocated: 120.02 MB
	Gen 0: 373, Gen 1: 27, Gen 2: 2

	Forcing GC... done!

	Allocated: 21.60 MB
	Gen 0: 374, Gen 1: 28, Gen 2: 3

I feel these results also show that this PR has no real impact on Requests/sec.